requirements.txt
streamlit~=1.18.0
tensorflow~=2.17.0
numpy~=1.23.4
Pillow~=9.2.0
tensorflow-hub>=0.12.0
altair~=4.2.0